feat: implementing react-native gRPC client
