+++ gnoboard/package.json
@@ -18,6 +18,7 @@
     "expo": "~49.0.7",
     "expo-splash-screen": "~0.20.5",
     "expo-status-bar": "~1.6.0",
+    "fast-text-encoding": "^1.0.6",
     "react": "18.2.0",
     "react-native": "0.72.4",
     "react-native-safe-area-context": "4.6.3",
