+++ gnoboard/src/grpc/client.ts
@@ -3,11 +3,11 @@ import { createXHRGrpcWebTransport } from "./transport";
 import { GnomobileService } from "@gno/api/rpc_connect";
 
 // Create a GnomobileService client
-export function createClient() {
+export function createClient(port: Number) {
   return createPromiseClient(
     GnomobileService,
     createXHRGrpcWebTransport({
-      baseUrl: "https://demo.connectrpc.com",
+      baseUrl: "http://127.0.0.1:" + port,
     }),
   );
 }
