+++ service/config.go
@@ -1,0 +1,257 @@
+package service
+
+import (
+	"os"
+
+	"go.uber.org/zap"
+)
+
+// Config describes a set of settings for a GnomobileService
+type Config struct {
+	Logger         *zap.Logger
+	Remote         string
+	ChainID        string
+	RootDir        string
+	TmpDir         string
+	TcpPort        int
+	UseUdsListener bool
+	UseTcpListener bool
+}
+
+type GnomobileOption func(cfg *Config) error
+
+// FallBackOption is a structure that permits to fallback to a default option if the option is not set.
+type FallBackOption struct {
+	fallback func(cfg *Config) bool
+	opt      GnomobileOption
+}
+
+// --- Logger options ---
+
+// WithLogger set the given logger.
+var WithLogger = func(l *zap.Logger) GnomobileOption {
+	return func(cfg *Config) error {
+		cfg.Logger = l
+		return nil
+	}
+}
+
+// WithDefaultLogger init a noop logger.
+var WithDefaultLogger GnomobileOption = func(cfg *Config) error {
+	logger, err := zap.NewDevelopment()
+	if err != nil {
+		return err
+	}
+
+	cfg.Logger = logger
+
+	return nil
+}
+
+var fallbackLogger = FallBackOption{
+	fallback: func(cfg *Config) bool { return cfg.Logger == nil },
+	opt:      WithDefaultLogger,
+}
+
+// WithFallbackLogger sets the logger if no logger is set.
+var WithFallbackLogger GnomobileOption = func(cfg *Config) error {
+	if fallbackLogger.fallback(cfg) {
+		return fallbackLogger.opt(cfg)
+	}
+	return nil
+}
+
+// --- Remote options ---
+
+// WithRemote sets the given remote node address.
+var WithRemote = func(remote string) GnomobileOption {
+	return func(cfg *Config) error {
+		cfg.Remote = remote
+		return nil
+	}
+}
+
+// WithDefaultRemote inits a default remote node address.
+var WithDefaultRemote GnomobileOption = func(cfg *Config) error {
+	cfg.Remote = "testnet.gno.berty.io:26657"
+	return nil
+}
+
+var fallbackRemote = FallBackOption{
+	fallback: func(cfg *Config) bool { return cfg.Remote == "" },
+	opt:      WithDefaultRemote,
+}
+
+// WithFallbackRemote sets the remote node address if no address is set.
+var WithFallbacRemote GnomobileOption = func(cfg *Config) error {
+	if fallbackRemote.fallback(cfg) {
+		return fallbackRemote.opt(cfg)
+	}
+	return nil
+}
+
+// --- ChainID options ---
+
+// WithChainID sets the given chain ID.
+var WithChainID = func(chainID string) GnomobileOption {
+	return func(cfg *Config) error {
+		cfg.ChainID = chainID
+		return nil
+	}
+}
+
+// WithDefaultChainID sets a default chain ID.
+var WithDefaultChainID GnomobileOption = func(cfg *Config) error {
+	cfg.ChainID = "dev"
+
+	return nil
+}
+
+var fallbackChainID = FallBackOption{
+	fallback: func(cfg *Config) bool { return cfg.ChainID == "" },
+	opt:      WithDefaultChainID,
+}
+
+// WithFallbackChainID sets the chain ID if no chain ID is set.
+var WithFallbacChainID GnomobileOption = func(cfg *Config) error {
+	if fallbackChainID.fallback(cfg) {
+		return fallbackChainID.opt(cfg)
+	}
+	return nil
+}
+
+// --- RootDir options ---
+
+// WithRootDir sets the given root directory path.
+var WithRootDir = func(rootDir string) GnomobileOption {
+	return func(cfg *Config) error {
+		cfg.RootDir = rootDir
+		return nil
+	}
+}
+
+// WithDefaultRootDir sets a default root directory in a temporary folder.
+var WithDefaultRootDir GnomobileOption = func(cfg *Config) error {
+	rootDir, err := os.MkdirTemp("", "gnomobile")
+	if err != nil {
+		return err
+	}
+
+	cfg.RootDir = rootDir
+
+	return nil
+}
+
+var fallbackRootDir = FallBackOption{
+	fallback: func(cfg *Config) bool { return cfg.RootDir == "" },
+	opt:      WithDefaultRootDir,
+}
+
+// WithFallbackRootDir sets the default root directory if no directory is set.
+var WithFallbackRootDir GnomobileOption = func(cfg *Config) error {
+	if fallbackRootDir.fallback(cfg) {
+		return fallbackRootDir.opt(cfg)
+	}
+	return nil
+}
+
+// --- tmpDir options ---
+
+// WithTmpDir sets the given temporary path.
+var WithTmpDir = func(path string) GnomobileOption {
+	return func(cfg *Config) error {
+		cfg.TmpDir = path
+		return nil
+	}
+}
+
+// WithDefaultTmpDir sets a default temporary path.
+var WithDefaultTmpDir GnomobileOption = func(cfg *Config) error {
+	// dependency
+	if err := WithFallbackRootDir(cfg); err != nil {
+		return err
+	}
+
+	cfg.TmpDir = cfg.RootDir
+
+	return nil
+}
+
+var fallbackTmpDir = FallBackOption{
+	fallback: func(cfg *Config) bool { return cfg.TmpDir == "" },
+	opt:      WithDefaultTmpDir,
+}
+
+// WithFallbackTmpDir sets the default temporary path if no path is set.
+var WithFallbackTmpDir GnomobileOption = func(cfg *Config) error {
+	if fallbackTmpDir.fallback(cfg) {
+		return fallbackTmpDir.opt(cfg)
+	}
+	return nil
+}
+
+// --- tcpPort options ---
+
+// WithTcpPort sets the given tcp port to serve the gRPC server.
+var WithTcpPort = func(port int) GnomobileOption {
+	return func(cfg *Config) error {
+		cfg.TcpPort = port
+		return nil
+	}
+}
+
+// --- useUdsListener options ---
+
+// WithUseUDSListener sets the gRPC server to serve on a Unix Domain Socket listener.
+var WithUseUdsListener = func() GnomobileOption {
+	return func(cfg *Config) error {
+		cfg.UseUdsListener = true
+		return nil
+	}
+}
+
+// --- useTcpListener options ---
+
+// WithUseTcpListener sets the gRPC server to serve on a TCP listener.
+var WithUseTcpListener = func() GnomobileOption {
+	return func(cfg *Config) error {
+		cfg.UseTcpListener = true
+		return nil
+	}
+}
+
+// --- listener fallback option ---
+
+// WithUseDefaultListener sets a TCP listener for the gRPC server.
+var WithUseDefaultListener GnomobileOption = func(cfg *Config) error {
+	cfg.UseTcpListener = true
+	return nil
+}
+
+var fallbackListener = FallBackOption{
+	fallback: func(cfg *Config) bool { return !cfg.UseTcpListener && !cfg.UseUdsListener },
+	opt:      WithUseDefaultListener,
+}
+
+// --- Fallback options ---
+
+var defaults = []FallBackOption{
+	fallbackLogger,
+	fallbackRemote,
+	fallbackChainID,
+	fallbackRootDir,
+	fallbackTmpDir,
+}
+
+// WithFallbackDefaults sets the default options if no option is set.
+var WithFallbackDefaults GnomobileOption = func(cfg *Config) error {
+	for _, def := range defaults {
+		if !def.fallback(cfg) {
+			continue
+		}
+		if err := def.opt(cfg); err != nil {
+			return err
+		}
+	}
+	return nil
+}
